Double to BigDecimal

+++ expense-service/src/main/java/com/appswella/wisepaise/model/Expense.java
@@ -5,10 +5,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.math.BigDecimal;
 import java.util.List;
-import java.util.Map;
+
 
 @Data
 @AllArgsConstructor
@@ -20,7 +18,7 @@ public class Expense {
     private String expenseId;
     private String expenseTitle;
     private String expenseNote;
-    private BigDecimal expenseAmount;
+    private Double expenseAmount;
     private String expenseSpendType;
     private String expenseDate;
     private String expenseCategory;
